requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.15
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
httpx==0.26.0
flake8==7.0.0